+++ tests/StringsTest.php
@@ -31,9 +31,11 @@ class StringsTest extends TestFixture
    }
    public function testToLocalizedFormattedDateString()
    {
-      setlocale(LC_ALL, 'fr_FR');
+      $cache = setlocale(LC_TIME, 0);
+      setlocale(LC_TIME, 'German');
       $d = Carbon::create(1975, 12, 25, 14, 15, 16);
-      $this->assertSame('Jeudi 25 décembre 1975', $d->formatLocalized('%A %d %B %Y'));
+      $this->assertSame('Donnerstag 25 Dezember 1975', $d->formatLocalized('%A %d %B %Y'));
+      setlocale(LC_TIME, $cache);
    }
    public function testToTimeString()
    {